backend/requirements.txt
asgiref==3.8.1
astroid==3.3.5
cffi==1.17.1
cryptography==43.0.3
dill==0.3.9
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
isort==5.13.2
mccabe==0.7.0
pillow==11.0.0
platformdirs==4.3.6
pycparser==2.22
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlparse==0.5.1
tomlkit==0.13.2
